Misc: solve circular import issue
