+++ app/services/openai_service.py
@@ -11,8 +11,8 @@ from openai.types.beta import Thread
 from app.utils.whatsapp_utils import get_text_payload
 from db.utils import check_if_thread_exists, store_message, store_thread
 from app.config import llm_settings
-from app.utils.whatsapp_utils import send_message
 from app.tools.generate_exercise.generate_exercise import exercise_generator
+from app.services.whatsapp_service import whatsapp_client
 
 
 class OpenAIClient:
@@ -137,7 +137,7 @@ class OpenAIClient:
         # Send a message indicating tool execution to the user.
         data = get_text_payload(wa_id, msg)
         store_message(wa_id, msg, role="twiga")
-        await send_message(data)
+        await whatsapp_client.send_message(data)
 
     async def _get_latest_assistant_message(self, thread_id: str) -> str:
         # Retrieve the most recent message generated by the assistant.
