+++ app/utils/logging_utils.py
@@ -4,7 +4,7 @@ import httpx
 logger = logging.getLogger(__name__)
 
 
-async def log_httpx_response(response: httpx.Response) -> None:
+def log_httpx_response(response: httpx.Response) -> None:
     # ANSI escape code mapping
     ANSI_COLOR_CODES = {
         "green": "32",
@@ -23,14 +23,17 @@ async def log_httpx_response(response: httpx.Response) -> None:
     # Color formatting based on the status code
     if 200 <= status_code < 300:
         status_color = ANSI_COLOR_CODES["green"]
+        log_method = logger.debug
     elif 300 <= status_code < 400:
         status_color = ANSI_COLOR_CODES["cyan"]
+        log_method = logger.info
     elif 400 <= status_code < 500:
         status_color = ANSI_COLOR_CODES["yellow"]
+        log_method = logger.warning
     else:
         status_color = ANSI_COLOR_CODES["red"]
+        log_method = logger.error
 
-    # Logging the details with color coding
-    logger.debug(f"Response URL: {url}")
-    logger.debug(f"Response Status: \033[{status_color}m{status_code}\033[0m")
-    logger.debug(f"Response Body: \033[35m{body}\033[0m")  # Magenta for body
+    log_method(f"Response URL: {url}")
+    log_method(f"Response Status: \033[{status_color}m{status_code}\033[0m")
+    log_method(f"Response Body: \033[35m{body}\033[0m")
